Refactor auth and DB initialization to fix TypeError
- Centralized database initialization in `backend/database.js`.
- Centralized authentication middleware (`authenticateToken`, `authorize`) in `backend/auth.js`, using the shared DB instance.
- Updated `backend/index.js` and all route files in `backend/routes/` to use the singleton DB and auth functions.
- This resolves a `TypeError: authorize is not a function` in `experiments.js` likely caused by duplicated initializations or circular dependencies.
- Corrected usage of `req.user.userId` (was `req.user.id`) in `instruments.js`.
- Removed duplicate definitions of auth middleware and DB connections.
