+++ backend/routes/samples.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const { db, authenticateToken, authorize } = require('../auth');
+const { authenticateToken, authorize } = require('../auth');
+const db = require('../database'); // Import shared db instance
 
 // Placeholder success response for non-implemented routes
 const placeholderHandler = (req, res) => {
